refactor!: remove trigger.before; pipelines replace preflight orchestration
Deletes the trigger.before preflight machinery now that kind: PipelineTask
(PR1-5 + #346) is the orchestration unit and the only two consumers
(buildin/relay-server, cloudflared docs example) have been migrated.

Removed from pkg/trigger:
  - engine.go: runPrereqs, dispatchPipelineStage, validateBeforeRefs,
    detectBeforeCycle, inputParamsRefRe, finalizePreflightFailure, the
    runOneShotPreflight branch in fireAsync, the notifyPrereqCompletion call
    in FireChain, and the validateBeforeRefs call in Register; startDaemon/
    startDaemonInternal collapsed to a single non-preflight startDaemon.
  - daemon_state.go: notifyPrereqCompletion, propagateBeforeRerun,
    beforeContains, and the DaemonPrereqRunning/DaemonPrereqFailed states.

Removed from pkg/task:
  - spec.go: TriggerConfig.Before field, BeforeEntry type +
    UnmarshalYAML/MarshalYAML, and the trigger.before validate() block.
  - template.go: the trigger.before override-expansion loop in expandSpec.
  - taskset/override.go: the Trigger.Before deep-clone in copySpec.

Kept (pipeline machinery depends on these): Overrides,
validatePerEdgeOverrides (still used by trigger.chain), InputContext,
ResolveInputOutputList/Map, all of pipeline_runner.go + registerpipeline.go
(firePipeline, dispatchStage, fireStageRaw, finish, handlePipelineStageRerun,
validatePipelineRefs, deferredPipelines, etc.) — and the
handlePipelineStageRerun call in FireChain (the pipeline analogue, PR4).
DaemonFailedAfterPreflight is kept: it now means a daemon body launch failed
(its name is historical). preflightEnv (secrets/env-resolution preflight) is
unrelated to trigger.before and untouched.

Tests: deleted the obsolete preflight test files
(engine_preflight_test.go, engine_oneshot_preflight_test.go,
e2e_oneshot_preflight_test.go, e2e_template_preflight_pipeline_test.go,
e2e_relay_template_pipeline_test.go). Migrated the still-valuable daemon
lifecycle coverage (DaemonState round-trip, daemon starts immediately,
daemon-launch-failure → DaemonFailedAfterPreflight, the
onDaemonRunFinished #325/#329/#332 transitions) into the new
engine_daemon_lifecycle_test.go; migrated the per-edge override expansion +
no-exfiltration assertions to expandPipeline tests in template_test.go;
kept the chain per-edge override test in engine_per_edge_overrides_test.go;
moved shared helpers (loadBuildinWriteLocalAs, pollFileContents) into
surviving files. TestLoadDir_BuildinTasksParse now loads via LoadKindedDir
so the relay-server PipelineTask parses.

BREAKING CHANGE: trigger.before has been removed. Daemon-with-preflight and
one-shot preflight pipelines must be expressed as a kind: PipelineTask whose
stages render config and whose terminal stage is the daemon/body task. The
field is rejected at load time (see the follow-up commit).
