--- pkg/trigger/engine_preflight_test.go
@@ -1,1970 +1,0 @@
-package trigger
-
-// Tests for the cross-spec validation pass that the engine runs at
-// registration time on tasks declaring `trigger.before`. Per-spec validation
-// (see pkg/task.Spec.validate) can only enforce shape: it cannot check that
-// the referenced task actually exists in the registry or that it is itself a
-// one-shot task rather than another daemon. Both rules live here.
-
-import (
-	"context"
-	"errors"
-	"fmt"
-	"reflect"
-	"strings"
-	"sync"
-	"sync/atomic"
-	"testing"
-	"time"
-
-	"github.com/dicode/dicode/pkg/db"
-	"github.com/dicode/dicode/pkg/registry"
-	pkgruntime "github.com/dicode/dicode/pkg/runtime"
-	"github.com/dicode/dicode/pkg/task"
-	"go.uber.org/zap"
-)
-
-func TestRegister_BeforeUnknownTaskRejected(t *testing.T) {
-	e := newTestEnv(t)
-	daemon := &task.Spec{
-		ID:      "d",
-		Name:    "d",
-		Runtime: task.RuntimeDocker,
-		Docker:  &task.DockerConfig{Image: "alpine"},
-		Trigger: task.TriggerConfig{Daemon: true, Before: []task.BeforeEntry{{Task: "missing"}}},
-		Enabled: true,
-	}
-	err := e.engine.Register(daemon)
-	if err == nil || !strings.Contains(err.Error(), "missing") {
-		t.Errorf("expected error referencing unknown task, got %v", err)
-	}
-}
-
-func TestRegister_BeforeDaemonRejected(t *testing.T) {
-	e := newTestEnv(t)
-	other := &task.Spec{
-		ID:      "other-daemon",
-		Name:    "other-daemon",
-		Runtime: task.RuntimeDocker,
-		Docker:  &task.DockerConfig{Image: "alpine"},
-		Trigger: task.TriggerConfig{Daemon: true},
-		Enabled: true,
-	}
-	if err := e.reg.Register(other); err != nil {
-		t.Fatalf("seed reg.Register: %v", err)
-	}
-	target := &task.Spec{
-		ID:      "d",
-		Name:    "d",
-		Runtime: task.RuntimeDocker,
-		Docker:  &task.DockerConfig{Image: "alpine"},
-		Trigger: task.TriggerConfig{Daemon: true, Before: []task.BeforeEntry{{Task: "other-daemon"}}},
-		Enabled: true,
-	}
-	err := e.engine.Register(target)
-	if err == nil || !strings.Contains(err.Error(), "daemon") {
-		t.Errorf("expected error rejecting daemon-as-before, got %v", err)
-	}
-}
-
-// Cycle case: rejected structurally by the per-spec + cross-spec rules.
-//
-// trigger.before is only valid on daemon tasks (per-spec) AND it cannot
-// reference a daemon (cross-spec). The only way a cycle could form is
-// through a prereq task having its own trigger.before pointing back at the
-// daemon — but only daemon tasks may have trigger.before. Therefore cycles
-// are unreachable, and an explicit cycle-detection test is structurally
-// impossible to write. The validator comment in validateBeforeRefs
-// captures this reasoning so future readers don't add a redundant check.
-
-// TestDaemonState_Default verifies that DaemonState() returns the zero
-// value (DaemonStopped) for unknown task IDs and for daemons that haven't
-// been started yet — operators viewing the WebUI before the engine has
-// fired the preflight should see "stopped", not a blank field.
-func TestDaemonState_Default(t *testing.T) {
-	e := newTestEnv(t)
-	if got := e.engine.DaemonState("unknown"); got != DaemonStopped {
-		t.Errorf("unknown daemon state = %q, want stopped", got)
-	}
-}
-
-// TestDaemonState_SetGet verifies that setDaemonState/DaemonState round-trip
-// the seven enum values without contention.
-func TestDaemonState_SetGet(t *testing.T) {
-	e := newTestEnv(t)
-	states := []DaemonState{
-		DaemonPrereqRunning,
-		DaemonPrereqFailed,
-		DaemonRunning,
-		DaemonStopping,
-		DaemonFailedAfterPreflight,
-		DaemonCrashed,
-		DaemonStopped,
-	}
-	for _, want := range states {
-		e.engine.setDaemonState("d", want)
-		if got := e.engine.DaemonState("d"); got != want {
-			t.Errorf("setDaemonState(%q) → DaemonState = %q, want %q", want, got, want)
-		}
-	}
-}
-
-// TestRegister_BeforeValid_NonDaemonTarget exercises the happy path: a
-// daemon with a `before:` list referencing a one-shot task that exists in
-// the registry registers without error.
-func TestRegister_BeforeValid_NonDaemonTarget(t *testing.T) {
-	e := newTestEnv(t)
-	prereq := &task.Spec{
-		ID:      "render",
-		Name:    "render",
-		Runtime: task.RuntimeDeno,
-		Trigger: task.TriggerConfig{Manual: true},
-		Enabled: true,
-	}
-	if err := e.reg.Register(prereq); err != nil {
-		t.Fatalf("seed reg.Register: %v", err)
-	}
-	daemon := &task.Spec{
-		ID:      "d",
-		Name:    "d",
-		Runtime: task.RuntimeDocker,
-		Docker:  &task.DockerConfig{Image: "alpine"},
-		Trigger: task.TriggerConfig{Daemon: true, Before: []task.BeforeEntry{{Task: "render"}}},
-		Enabled: true,
-	}
-	if err := e.engine.Register(daemon); err != nil {
-		t.Errorf("unexpected error on valid before-list: %v", err)
-	}
-}
-
-// preflightExec is a controllable Executor for the preflight tests. Each
-// task is keyed by ID; the configured fn runs synchronously and decides
-// the final status. Daemons never finish unless explicitly released.
-type preflightExec struct {
-	reg *registry.Registry
-
-	mu       sync.Mutex
-	fns      map[string]func(taskID, runID string) string
-	returns  map[string]func(taskID, runID string, spec *task.Spec) interface{}
-	captured map[string]*task.Spec    // runID → spec snapshot (for assertions on merged Params)
-	daemon   map[string]chan struct{} // daemon taskID → block channel (closed = exit)
-	runs     atomic.Int32             // count of executed runs (any task)
-	failing  map[string]bool          // task IDs whose Execute should always return StatusFailure
-}
-
-func newPreflightExec(reg *registry.Registry) *preflightExec {
-	return &preflightExec{
-		reg:      reg,
-		fns:      make(map[string]func(string, string) string),
-		returns:  make(map[string]func(string, string, *task.Spec) interface{}),
-		captured: make(map[string]*task.Spec),
-		daemon:   make(map[string]chan struct{}),
-		failing:  make(map[string]bool),
-	}
-}
-
-func (p *preflightExec) setFn(taskID string, fn func(taskID, runID string) string) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	p.fns[taskID] = fn
-}
-
-// setReturnFn registers a function that produces a return value for taskID's
-// Execute call. The closure receives the spec snapshot the executor was
-// handed, so tests can read post-override params (e.g. the resolved
-// ${input.output} default) and decide what to return. Status is always
-// StatusSuccess for returns-set tasks unless setFn also overrides it.
-func (p *preflightExec) setReturnFn(taskID string, fn func(taskID, runID string, spec *task.Spec) interface{}) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	p.returns[taskID] = fn
-}
-
-// markFailing pins taskID's Execute to always return StatusFailure, so
-// pipeline-failure tests don't need to clone setFn boilerplate.
-func (p *preflightExec) markFailing(taskID string) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	p.failing[taskID] = true
-}
-
-// specForRun returns the spec snapshot Execute captured for the given
-// runID, or nil if no run was recorded under that ID.
-func (p *preflightExec) specForRun(runID string) *task.Spec {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	return p.captured[runID]
-}
-
-// markDaemon flags a task ID as a long-running daemon. Subsequent Execute
-// calls for that task block until the run's context is cancelled (i.e.
-// until KillRun is invoked or the engine shuts down).
-func (p *preflightExec) markDaemon(taskID string) {
-	p.mu.Lock()
-	if p.daemon[taskID] == nil {
-		p.daemon[taskID] = make(chan struct{})
-	}
-	p.mu.Unlock()
-}
-
-func (p *preflightExec) Execute(ctx context.Context, spec *task.Spec, opts pkgruntime.RunOptions) (*pkgruntime.RunResult, error) {
-	p.runs.Add(1)
-	p.mu.Lock()
-	daemonCh := p.daemon[spec.ID]
-	fn := p.fns[spec.ID]
-	retFn := p.returns[spec.ID]
-	failing := p.failing[spec.ID]
-	// Capture a shallow copy of the spec so tests can assert on merged
-	// Params (e.g. ${input.output} resolution) without racing later
-	// engine mutations. Params is a slice; copy it explicitly so a later
-	// override on the registry spec wouldn't disturb the snapshot.
-	specCopy := *spec
-	if len(spec.Params) > 0 {
-		specCopy.Params = append(task.Params(nil), spec.Params...)
-	}
-	p.captured[opts.RunID] = &specCopy
-	p.mu.Unlock()
-
-	if daemonCh != nil {
-		// Long-running daemon: block until the run's context is cancelled
-		// (KillRun path). FinishRun(Success) before returning so onDaemon-
-		// RunFinished doesn't treat the daemon as crashed. Mirrors how a
-		// real daemon exits gracefully on shutdown.
-		<-ctx.Done()
-		_ = p.reg.FinishRun(context.Background(), opts.RunID, registry.StatusSuccess)
-		return &pkgruntime.RunResult{}, nil
-	}
-	status := registry.StatusSuccess
-	if failing {
-		status = registry.StatusFailure
-	}
-	if fn != nil {
-		status = fn(spec.ID, opts.RunID)
-	}
-	if status == registry.StatusFailure {
-		_ = p.reg.FinishRunWithReason(context.Background(), opts.RunID, status, "test-injected: failure")
-		return &pkgruntime.RunResult{}, errors.New("injected failure")
-	}
-	res := &pkgruntime.RunResult{}
-	if retFn != nil {
-		v := retFn(spec.ID, opts.RunID, &specCopy)
-		res.ReturnValue = v
-		// Mirror the Deno runtime's default (runtime/deno/runtime.go:
-		// "r.ChainInput = result.ReturnValue" when no structured
-		// output is present). FireChain — and the mid-pipeline re-fire
-		// propagator — reads ChainInput when an upstream completes; we
-		// want tests to see the same value flow.
-		res.ChainInput = v
-	}
-	_ = p.reg.FinishRun(context.Background(), opts.RunID, status)
-	return res, nil
-}
-
-// newPreflightEnv returns a test engine + reg + controllable executor for
-// the preflight gating tests. Tasks complete synchronously (or block, for
-// marked daemons) without spawning Deno or Docker — we register the same
-// preflightExec for both the deno and docker runtimes so daemon specs
-// (which are typically runtime: docker for container daemons) dispatch
-// without needing the real docker manager.
-func newPreflightEnv(t *testing.T) (*Engine, *registry.Registry, *preflightExec) {
-	t.Helper()
-	d, err := db.Open(db.Config{Type: "sqlite", Path: ":memory:"})
-	if err != nil {
-		t.Fatalf("open db: %v", err)
-	}
-	t.Cleanup(func() { d.Close() })
-	reg := registry.New(d)
-	exec := newPreflightExec(reg)
-	eng := New(reg, exec, zap.NewNop())
-	eng.RegisterExecutor(task.RuntimeDocker, exec)
-	return eng, reg, exec
-}
-
-func waitUntil(t *testing.T, timeout time.Duration, cond func() bool, msg string) {
-	t.Helper()
-	deadline := time.Now().Add(timeout)
-	for time.Now().Before(deadline) {
-		if cond() {
-			return
-		}
-		time.Sleep(5 * time.Millisecond)
-	}
-	t.Fatalf("timed out waiting: %s", msg)
-}
-
-// hasSuccessfulRunFor reports whether taskID has at least one run row in
-// the registry with the given status.
-func hasRunWithStatus(t *testing.T, reg *registry.Registry, taskID, status string) bool {
-	t.Helper()
-	runs, err := reg.ListRuns(context.Background(), taskID, 5)
-	if err != nil {
-		t.Fatalf("ListRuns: %v", err)
-	}
-	for _, r := range runs {
-		if r.Status == status {
-			return true
-		}
-	}
-	return false
-}
-
-// TestPreflight_DaemonWaitsForPrereqSuccess verifies the gating contract:
-// the daemon's run does not start until every task in its before-list has
-// finished with status=success. Asserted by ordering — prereq must have a
-// success run before the daemon's run shows up.
-func TestPreflight_DaemonWaitsForPrereqSuccess(t *testing.T) {
-	eng, reg, exec := newPreflightEnv(t)
-
-	// Prereq: completes successfully.
-	prereq := &task.Spec{
-		ID:      "render",
-		Name:    "render",
-		Runtime: task.RuntimeDeno,
-		Trigger: task.TriggerConfig{Manual: true},
-		Enabled: true,
-	}
-	if err := reg.Register(prereq); err != nil {
-		t.Fatalf("reg.Register prereq: %v", err)
-	}
-
-	// Daemon: blocks until released. restart:never so the executor's
-	// eventual exit doesn't loop the test.
-	daemon := &task.Spec{
-		ID:      "d",
-		Name:    "d",
-		Runtime: task.RuntimeDocker,
-		Docker:  &task.DockerConfig{Image: "alpine"},
-		Trigger: task.TriggerConfig{Daemon: true, Before: []task.BeforeEntry{{Task: "render"}}, Restart: "never"},
-		Enabled: true,
-	}
-	if err := reg.Register(daemon); err != nil {
-		t.Fatalf("reg.Register daemon: %v", err)
-	}
-	exec.markDaemon("d")
-
-	if err := eng.Register(daemon); err != nil {
-		t.Fatalf("eng.Register daemon: %v", err)
-	}
-
-	// Prereq must complete to success.
-	waitUntil(t, 5*time.Second, func() bool {
-		return hasRunWithStatus(t, reg, "render", registry.StatusSuccess)
-	}, "prereq render never succeeded")
-
-	// Daemon must transition to Running once preflight succeeds.
-	waitUntil(t, 5*time.Second, func() bool {
-		return eng.DaemonState("d") == DaemonRunning
-	}, "daemon never reached Running")
-}
-
-// TestPreflight_NoBefore_StartsImmediately verifies that a daemon with an
-// empty before-list does not pay any preflight overhead: it goes straight
-// to Running without the engine touching daemonStates' PrereqRunning state.
-func TestPreflight_NoBefore_StartsImmediately(t *testing.T) {
-	eng, reg, exec := newPreflightEnv(t)
-	daemon := &task.Spec{
-		ID:      "d-noprereq",
-		Name:    "d-noprereq",
-		Runtime: task.RuntimeDocker,
-		Docker:  &task.DockerConfig{Image: "alpine"},
-		Trigger: task.TriggerConfig{Daemon: true, Restart: "never"}, // no Before
-		Enabled: true,
-	}
-	if err := reg.Register(daemon); err != nil {
-		t.Fatalf("reg.Register: %v", err)
-	}
-	exec.markDaemon("d-noprereq")
-
-	if err := eng.Register(daemon); err != nil {
-		t.Fatalf("eng.Register: %v", err)
-	}
-	waitUntil(t, 5*time.Second, func() bool {
-		return eng.DaemonState("d-noprereq") == DaemonRunning
-	}, "daemon-without-before never reached Running")
-}
-
-// makeDaemonSpec is a tiny convenience to keep the table-style tests below
-// from drowning in struct literals. Kept local to this file.
-func makeDaemonSpec(id string, before []string) *task.Spec {
-	entries := make([]task.BeforeEntry, len(before))
-	for i, t := range before {
-		entries[i] = task.BeforeEntry{Task: t}
-	}
-	return &task.Spec{
-		ID:      id,
-		Name:    id,
-		Runtime: task.RuntimeDocker,
-		Docker:  &task.DockerConfig{Image: "alpine"},
-		// restart: never so a clean executor exit doesn't trigger the
-		// existing "always restart" hook AND the prereq-driven restart
-		// path together (would double-count daemon runs in the
-		// coalescing test).
-		Trigger: task.TriggerConfig{Daemon: true, Before: entries, Restart: "never"},
-		Enabled: true,
-	}
-}
-
-func makeOneShotSpec(id string) *task.Spec {
-	return &task.Spec{
-		ID:      id,
-		Name:    id,
-		Runtime: task.RuntimeDeno,
-		Trigger: task.TriggerConfig{Manual: true},
-		Enabled: true,
-	}
-}
-
-// TestPreflight_DaemonRestartsOnPrereqRerun verifies that re-running a
-// prereq task after the daemon is up triggers a restart: the daemon's
-// current run is killed, preflight re-runs, and the daemon comes back up.
-//
-// Asserted by counting daemon-task runs in the registry — must transition
-// from 1 to 2 after the prereq is re-fired.
-func TestPreflight_DaemonRestartsOnPrereqRerun(t *testing.T) {
-	eng, reg, exec := newPreflightEnv(t)
-
-	prereq := makeOneShotSpec("render")
-	if err := reg.Register(prereq); err != nil {
-		t.Fatal(err)
-	}
-
-	daemon := makeDaemonSpec("d", []string{"render"})
-	if err := reg.Register(daemon); err != nil {
-		t.Fatal(err)
-	}
-	// markDaemon makes the executor block until the run's context is
-	// cancelled. KillRun (invoked by queueDaemonRestart) is the cancel
-	// trigger.
-	exec.markDaemon("d")
-
-	if err := eng.Register(daemon); err != nil {
-		t.Fatalf("eng.Register: %v", err)
-	}
-
-	// Wait for daemon's first run to be active.
-	waitUntil(t, 5*time.Second, func() bool {
-		return eng.DaemonState("d") == DaemonRunning
-	}, "daemon never reached Running for first start")
-
-	// Re-fire the prereq. The success completion notifies the engine,
-	// which should KillRun the current daemon run, re-run preflight, and
-	// start a fresh daemon run.
-	if _, err := eng.FireManual(context.Background(), "render", nil); err != nil {
-		t.Fatalf("FireManual prereq: %v", err)
-	}
-
-	// Wait for a second daemon run to appear AND reach Running.
-	waitUntil(t, 10*time.Second, func() bool {
-		runs, _ := reg.ListRuns(context.Background(), "d", 10)
-		var daemonRuns int
-		for _, r := range runs {
-			if r.TriggerSource == registry.TriggerDaemon {
-				daemonRuns++
-			}
-		}
-		return daemonRuns >= 2 && eng.DaemonState("d") == DaemonRunning
-	}, "daemon never restarted after prereq re-run")
-}
-
-// TestPreflight_RestartCoalesces verifies that two rapid prereq re-runs
-// produce AT MOST one extra daemon restart — the second prereq completion
-// arriving while a restart is already in flight is coalesced. Without
-// coalescing, busy prereq schedules (e.g. credential rotators on a short
-// cron) would thrash the daemon.
-func TestPreflight_RestartCoalesces(t *testing.T) {
-	eng, reg, exec := newPreflightEnv(t)
-
-	prereq := makeOneShotSpec("render")
-	if err := reg.Register(prereq); err != nil {
-		t.Fatal(err)
-	}
-	daemon := makeDaemonSpec("d", []string{"render"})
-	if err := reg.Register(daemon); err != nil {
-		t.Fatal(err)
-	}
-	exec.markDaemon("d")
-
-	if err := eng.Register(daemon); err != nil {
-		t.Fatalf("eng.Register: %v", err)
-	}
-	waitUntil(t, 5*time.Second, func() bool {
-		return eng.DaemonState("d") == DaemonRunning
-	}, "first start")
-
-	// Fire prereq twice in rapid succession. The second fire must be
-	// coalesced into the first restart (restartGate.tryAcquire returns
-	// false on the second attempt while the first is still in flight).
-	for i := 0; i < 2; i++ {
-		if _, err := eng.FireManual(context.Background(), "render", nil); err != nil {
-			t.Fatalf("FireManual prereq: %v", err)
-		}
-	}
-
-	// After the restart settles, total daemon runs should be 2 — first
-	// boot plus one coalesced restart. Wait briefly to allow any erroneous
-	// second restart to manifest.
-	waitUntil(t, 10*time.Second, func() bool {
-		runs, _ := reg.ListRuns(context.Background(), "d", 10)
-		var daemonRuns int
-		for _, r := range runs {
-			if r.TriggerSource == registry.TriggerDaemon {
-				daemonRuns++
-			}
-		}
-		return daemonRuns >= 2 && eng.DaemonState("d") == DaemonRunning
-	}, "restart never completed")
-	time.Sleep(500 * time.Millisecond)
-
-	runs, _ := reg.ListRuns(context.Background(), "d", 10)
-	var daemonRuns int
-	for _, r := range runs {
-		if r.TriggerSource == registry.TriggerDaemon {
-			daemonRuns++
-		}
-	}
-	if daemonRuns > 2 {
-		t.Errorf("expected at most 2 daemon runs (first boot + one coalesced restart), got %d", daemonRuns)
-	}
-}
-
-// TestPreflight_PrereqFailureBlocksFirstStart verifies that a daemon
-// whose preflight fails on the very first attempt:
-//
-//   - does NOT have its daemon-task fired,
-//   - ends up in state=DaemonPrereqFailed (operator-facing diagnosis).
-//
-// The default 'restart: always' would otherwise mask the failure by
-// retrying forever — preflight failure is a config-level error and
-// should surface, not loop.
-func TestPreflight_PrereqFailureBlocksFirstStart(t *testing.T) {
-	eng, reg, exec := newPreflightEnv(t)
-
-	prereq := makeOneShotSpec("render")
-	if err := reg.Register(prereq); err != nil {
-		t.Fatal(err)
-	}
-	// Make render fail.
-	exec.setFn("render", func(string, string) string { return registry.StatusFailure })
-
-	daemon := makeDaemonSpec("d", []string{"render"})
-	if err := reg.Register(daemon); err != nil {
-		t.Fatal(err)
-	}
-	exec.markDaemon("d")
-
-	if err := eng.Register(daemon); err != nil {
-		t.Fatalf("eng.Register: %v", err)
-	}
-
-	// Wait until the engine has finished its preflight attempt and
-	// settled into PrereqFailed.
-	waitUntil(t, 5*time.Second, func() bool {
-		return eng.DaemonState("d") == DaemonPrereqFailed
-	}, "daemon never reached PrereqFailed")
-
-	// No daemon run should have been recorded.
-	runs, _ := reg.ListRuns(context.Background(), "d", 10)
-	for _, r := range runs {
-		if r.TriggerSource == registry.TriggerDaemon {
-			t.Errorf("daemon run %s was fired despite failed prereq", r.ID)
-		}
-	}
-}
-
-// TestPreflight_PrereqFailureLeavesRunningDaemonAlone verifies that a
-// prereq failure that happens AFTER the daemon is already up does not
-// disturb the running daemon. Rationale: an in-flight credential rotator
-// that hits a transient API error shouldn't take down a long-running
-// service. The daemon keeps using the last-known-good config; the
-// failure is visible in the prereq's run log.
-func TestPreflight_PrereqFailureLeavesRunningDaemonAlone(t *testing.T) {
-	eng, reg, exec := newPreflightEnv(t)
-
-	prereq := makeOneShotSpec("render")
-	if err := reg.Register(prereq); err != nil {
-		t.Fatal(err)
-	}
-	// First attempt succeeds (so the daemon comes up). We'll flip to
-	// failure after the daemon is Running.
-	exec.setFn("render", func(string, string) string { return registry.StatusSuccess })
-
-	daemon := makeDaemonSpec("d", []string{"render"})
-	if err := reg.Register(daemon); err != nil {
-		t.Fatal(err)
-	}
-	exec.markDaemon("d")
-
-	if err := eng.Register(daemon); err != nil {
-		t.Fatalf("eng.Register: %v", err)
-	}
-	waitUntil(t, 5*time.Second, func() bool {
-		return eng.DaemonState("d") == DaemonRunning
-	}, "daemon never came up")
-
-	// Capture the current daemon run ID. The same run should still own
-	// the registry slot when the test ends.
-	eng.daemonMu.Lock()
-	originalRunID := eng.daemonRuns["d"]
-	eng.daemonMu.Unlock()
-	if originalRunID == "" {
-		t.Fatal("daemon has no recorded run ID after Running state")
-	}
-
-	// Flip render to failure and re-fire it.
-	exec.setFn("render", func(string, string) string { return registry.StatusFailure })
-	if _, err := eng.FireManual(context.Background(), "render", nil); err != nil {
-		t.Fatalf("FireManual: %v", err)
-	}
-
-	// Give the engine time to (incorrectly) attempt a restart. A 1s
-	// settle window is generous given the engine doesn't sleep between
-	// FireChain and queueDaemonRestart.
-	time.Sleep(1 * time.Second)
-
-	if got := eng.DaemonState("d"); got != DaemonRunning {
-		t.Errorf("daemon state after prereq failure = %q, want running", got)
-	}
-	eng.daemonMu.Lock()
-	currentRunID := eng.daemonRuns["d"]
-	eng.daemonMu.Unlock()
-	if currentRunID != originalRunID {
-		t.Errorf("daemon run ID changed: was %q, now %q (the engine restarted on a failed prereq)", originalRunID, currentRunID)
-	}
-}
-
-var _ = fmt.Sprintf // keep fmt import live for future debug-helpers
-
-// TestPreflight_RegisterDaemon_Concurrent_NoDoubleStart pins finding #1 from
-// the PR-300 review: two concurrent registerDaemon calls for the same
-// preflight daemon must NOT each spawn a goroutine that fires preflight and
-// starts the daemon. Without coalescing, both callers observe `daemonRuns`
-// empty (since neither has populated it yet — the goroutine only assigns
-// `daemonRuns[id] = runID` AFTER preflight + fireAsync) and both run a
-// preflight chain + a daemon run, racing for the registry slot.
-//
-// The fix: gate the initial-start path with the same per-task lock used for
-// restart coalescing. The second concurrent call should be a no-op.
-//
-// We inject a barrier into the prereq executor so both registerDaemon calls
-// can reach `go startDaemon` before either preflight returns. After the
-// barrier releases, both goroutines (if the bug exists) would proceed to
-// fireAsync; with the fix, only one does.
-func TestPreflight_RegisterDaemon_Concurrent_NoDoubleStart(t *testing.T) {
-	eng, reg, exec := newPreflightEnv(t)
-
-	prereq := makeOneShotSpec("render")
-	if err := reg.Register(prereq); err != nil {
-		t.Fatal(err)
-	}
-
-	// Block the prereq until released so both registerDaemon callers are
-	// guaranteed to have spawned their goroutines before either preflight
-	// completes. Without this, the test would race against goroutine
-	// scheduling and the bug could slip past on a fast machine.
-	prereqRelease := make(chan struct{})
-	var prereqRuns atomic.Int32
-	exec.setFn("render", func(string, string) string {
-		prereqRuns.Add(1)
-		<-prereqRelease
-		return registry.StatusSuccess
-	})
-
-	daemon := makeDaemonSpec("d", []string{"render"})
-	if err := reg.Register(daemon); err != nil {
-		t.Fatal(err)
-	}
-	exec.markDaemon("d")
-
-	// Fire two concurrent Register calls. Use a barrier so they enter
-	// registerDaemon in true overlap rather than serially.
-	var startBarrier sync.WaitGroup
-	startBarrier.Add(1)
-	var done sync.WaitGroup
-	done.Add(2)
-	for i := 0; i < 2; i++ {
-		go func() {
-			defer done.Done()
-			startBarrier.Wait()
-			if err := eng.Register(daemon); err != nil {
-				t.Errorf("eng.Register: %v", err)
-			}
-		}()
-	}
-	startBarrier.Done()
-	done.Wait()
-
-	// Wait until preflight has actually started for at least one caller.
-	// (Both should reach the barrier; we only need one to be sure both
-	// registerDaemon calls have completed their spawn step.)
-	waitUntil(t, 2*time.Second, func() bool {
-		return prereqRuns.Load() >= 1
-	}, "no prereq run started after concurrent Register calls")
-
-	// Release the prereq barrier so any in-flight preflight goroutine(s)
-	// proceed to fireAsync.
-	close(prereqRelease)
-
-	// Wait for the daemon to reach Running so any double-start would have
-	// surfaced by now.
-	waitUntil(t, 5*time.Second, func() bool {
-		return eng.DaemonState("d") == DaemonRunning
-	}, "daemon never reached Running")
-	// Extra settle window: a second goroutine's fireAsync may land after
-	// the first daemon run reached Running. Without this, the bug could
-	// slip past the assertion below.
-	time.Sleep(500 * time.Millisecond)
-
-	// Assertion 1: exactly one preflight run.
-	if got := prereqRuns.Load(); got != 1 {
-		t.Errorf("expected exactly 1 preflight run, got %d (double-start race)", got)
-	}
-
-	// Assertion 2: exactly one daemon run in the registry.
-	runs, err := reg.ListRuns(context.Background(), "d", 10)
-	if err != nil {
-		t.Fatalf("ListRuns: %v", err)
-	}
-	var daemonRuns int
-	for _, r := range runs {
-		if r.TriggerSource == registry.TriggerDaemon {
-			daemonRuns++
-		}
-	}
-	if daemonRuns != 1 {
-		t.Errorf("expected exactly 1 daemon run, got %d (double-start race)", daemonRuns)
-	}
-}
-
-// TestRunPrereqs_RejectsInputOutputAtDispatch_LatentPR3 verifies the
-// runtime backstop for the ${input.output} contract. Registration
-// allows the token on non-first `before:` stages (PR3 will pipe the
-// previous stage's return value through), but PR2's runPrereqs hook
-// hardcodes upstreamOutput="" — so any token reference on a non-first
-// stage surfaces as ErrInputUnavailable at dispatch and the daemon
-// settles into DaemonPrereqFailed.
-//
-// When PR3 lands and runPrereqs threads real upstream output, this
-// test should be rewritten to assert successful interpolation — until
-// then, the loud failure is the desired contract: a literal token
-// must NOT reach the downstream.
-func TestRunPrereqs_RejectsInputOutputAtDispatch_LatentPR3(t *testing.T) {
-	eng, reg, _ := newPreflightEnv(t)
-
-	// Two one-shot prereqs. stage-a has no token; stage-b carries
-	// ${input.output} on its override, which is allowed at registration
-	// because it's not on before[0].
-	for _, id := range []string{"stage-a", "stage-b"} {
-		spec := &task.Spec{
-			ID:      id,
-			Name:    id,
-			Runtime: task.RuntimeDeno,
-			Trigger: task.TriggerConfig{Manual: true},
-			Enabled: true,
-		}
-		if err := reg.Register(spec); err != nil {
-			t.Fatalf("reg.Register %s: %v", id, err)
-		}
-	}
-
-	daemon := &task.Spec{
-		ID:      "daemon",
-		Name:    "daemon",
-		Runtime: task.RuntimeDocker,
-		Docker:  &task.DockerConfig{Image: "alpine"},
-		Trigger: task.TriggerConfig{
-			Daemon: true,
-			Before: []task.BeforeEntry{
-				{Task: "stage-a"}, // first stage, no token
-				{
-					Task: "stage-b",
-					Overrides: &task.Overrides{
-						Params: task.ParamOverrides{
-							{Name: "content", Default: "${input.output}"},
-						},
-					},
-				},
-			},
-			Restart: "never",
-		},
-		Enabled: true,
-	}
-	if err := reg.Register(daemon); err != nil {
-		t.Fatalf("reg.Register daemon: %v", err)
-	}
-	if err := eng.Register(daemon); err != nil {
-		t.Fatalf("eng.Register daemon: %v", err)
-	}
-
-	// runPrereqs fires both `before:` entries in parallel. stage-b's
-	// dispatch invokes ResolveInputOutputList with upstreamOutput="",
-	// returning ErrInputUnavailable; runPrereqs surfaces the error and
-	// the daemon ends up in DaemonPrereqFailed.
-	waitUntil(t, 5*time.Second, func() bool {
-		return eng.DaemonState("daemon") == DaemonPrereqFailed
-	}, "daemon never reached PrereqFailed for non-first stage token")
-}
-
-// TestRegister_RejectsInputOutputOnFirstBeforeStage verifies that the
-// engine refuses to register a daemon whose trigger.before[0] override
-// references ${input.output}. The first stage of a preflight pipeline
-// has no upstream return value, so the reference is statically
-// unresolvable — surface it at config-load time rather than at the
-// first daemon dispatch.
-func TestRegister_RejectsInputOutputOnFirstBeforeStage(t *testing.T) {
-	eng, reg, _ := newPreflightEnv(t)
-
-	upstream := &task.Spec{
-		ID:      "render",
-		Name:    "render",
-		Runtime: task.RuntimeDeno,
-		Trigger: task.TriggerConfig{Manual: true},
-		Enabled: true,
-	}
-	if err := reg.Register(upstream); err != nil {
-		t.Fatalf("register upstream: %v", err)
-	}
-
-	daemon := &task.Spec{
-		ID:      "my-daemon",
-		Name:    "my-daemon",
-		Runtime: task.RuntimeDocker,
-		Docker:  &task.DockerConfig{Image: "alpine"},
-		Trigger: task.TriggerConfig{
-			Daemon: true,
-			Before: []task.BeforeEntry{
-				{
-					Task: "render",
-					Overrides: &task.Overrides{
-						Params: task.ParamOverrides{
-							{Name: "content", Default: "${input.output}"},
-						},
-					},
-				},
-			},
-		},
-		Enabled: true,
-	}
-	err := eng.Register(daemon)
-	if err == nil {
-		t.Fatal("expected register to reject ${input.output} on before[0]")
-	}
-	if !strings.Contains(err.Error(), "${input.") {
-		t.Errorf("error should mention ${input. prefix; got %v", err)
-	}
-	if !strings.Contains(err.Error(), "before[0]") {
-		t.Errorf("error should pinpoint the offending stage; got %v", err)
-	}
-}
-
-// TestRegister_RejectsInputParamsOnPreflightStage verifies that
-// validateBeforeRefs rejects ${input.params.X} on any preflight
-// before-edge, not just before[0]. dispatchPipelineStage fires each
-// preflight stage with an empty RunOptions, so the upstream params
-// snapshot it threads forward is always nil — a reference would always
-// fail loudly at dispatch with ErrInputUnavailable. Surfacing it at
-// registration produces the clearer config-load error.
-func TestRegister_RejectsInputParamsOnPreflightStage(t *testing.T) {
-	eng, reg, _ := newPreflightEnv(t)
-
-	render := &task.Spec{
-		ID:      "render",
-		Name:    "render",
-		Runtime: task.RuntimeDeno,
-		Trigger: task.TriggerConfig{Manual: true},
-		Enabled: true,
-	}
-	if err := reg.Register(render); err != nil {
-		t.Fatalf("register render: %v", err)
-	}
-	write := &task.Spec{
-		ID:      "write",
-		Name:    "write",
-		Runtime: task.RuntimeDeno,
-		Trigger: task.TriggerConfig{Manual: true},
-		Enabled: true,
-	}
-	if err := reg.Register(write); err != nil {
-		t.Fatalf("register write: %v", err)
-	}
-
-	daemon := &task.Spec{
-		ID:      "my-daemon",
-		Name:    "my-daemon",
-		Runtime: task.RuntimeDocker,
-		Docker:  &task.DockerConfig{Image: "alpine"},
-		Trigger: task.TriggerConfig{
-			Daemon: true,
-			Before: []task.BeforeEntry{
-				// before[0]: no ${input.…} — must be a valid first stage
-				{Task: "render"},
-				// before[1]: references upstream params, which preflight
-				// stages never receive. Must be rejected at registration.
-				{
-					Task: "write",
-					Overrides: &task.Overrides{
-						Params: task.ParamOverrides{
-							{Name: "endpoint", Default: "${input.params.url}"},
-						},
-					},
-				},
-			},
-		},
-		Enabled: true,
-	}
-	err := eng.Register(daemon)
-	if err == nil {
-		t.Fatal("expected register to reject ${input.params.X} on a preflight before-edge")
-	}
-	if !strings.Contains(err.Error(), "${input.params") {
-		t.Errorf("error should mention ${input.params; got %v", err)
-	}
-	if !strings.Contains(err.Error(), "before[1]") {
-		t.Errorf("error should pinpoint the offending stage; got %v", err)
-	}
-	// ${input.output.X} on the same edge must continue to work —
-	// only the params channel is statically unresolvable on preflight.
-	daemon.Trigger.Before[1].Overrides.Params[0].Default = "${input.output.path}"
-	if err := eng.Register(daemon); err != nil {
-		t.Fatalf("${input.output.X} on before[i>0] should still register: %v", err)
-	}
-}
-
-// waitForCond polls cond until it returns true or the deadline expires.
-// Distinct from waitUntil (which fails the test on timeout): used by the
-// new sequential-pipeline tests where the caller wants to assert on the
-// final order/state after the cond reaches true, deferring failure to the
-// post-cond assertion.
-func waitForCond(timeout time.Duration, cond func() bool) bool {
-	deadline := time.Now().Add(timeout)
-	for time.Now().Before(deadline) {
-		if cond() {
-			return true
-		}
-		time.Sleep(5 * time.Millisecond)
-	}
-	return cond()
-}
-
-// TestBefore_SequentialExecution verifies the PR3 contract: every entry
-// in trigger.before runs in declaration order, each waiting for the
-// previous stage to reach terminal success. A parallel implementation
-// would interleave the three stages because each fn sleeps 50ms; a
-// sequential implementation produces the exact order [a, b, c].
-func TestBefore_SequentialExecution(t *testing.T) {
-	eng, reg, exec := newPreflightEnv(t)
-
-	var order []string
-	var orderMu sync.Mutex
-	recordRun := func(id string) func(_, _ string) string {
-		return func(_, _ string) string {
-			orderMu.Lock()
-			order = append(order, id)
-			orderMu.Unlock()
-			// Sleep so a parallel impl would clearly interleave.
-			time.Sleep(50 * time.Millisecond)
-			return registry.StatusSuccess
-		}
-	}
-	for _, id := range []string{"stage-a", "stage-b", "stage-c"} {
-		spec := makeOneShotSpec(id)
-		if err := reg.Register(spec); err != nil {
-			t.Fatalf("register %s: %v", id, err)
-		}
-		exec.setFn(id, recordRun(id))
-	}
-
-	daemon := makeDaemonSpec("d", []string{"stage-a", "stage-b", "stage-c"})
-	if err := reg.Register(daemon); err != nil {
-		t.Fatalf("register daemon: %v", err)
-	}
-	exec.markDaemon("d")
-
-	if err := eng.Register(daemon); err != nil {
-		t.Fatalf("eng.Register: %v", err)
-	}
-
-	waitUntil(t, 5*time.Second, func() bool {
-		orderMu.Lock()
-		defer orderMu.Unlock()
-		return len(order) == 3
-	}, "all three stages never completed")
-	orderMu.Lock()
-	defer orderMu.Unlock()
-	want := []string{"stage-a", "stage-b", "stage-c"}
-	if !reflect.DeepEqual(order, want) {
-		t.Errorf("before-list ran out of order: got %v, want %v", order, want)
-	}
-}
-
-// TestBefore_PipesInputOutputFieldThroughStages exercises the
-// post-#316 grammar extension on the preflight edge:
-// ${input.output.<field>} lifts a named string field out of a
-// structured upstream return value. Stage 1 returns
-// {path: "/var/run/relay.yaml"}; stage 2's override.content references
-// the field directly. The captured spec's content default must equal
-// the field value — the resolver type-asserts the map and pulls the
-// string out before the writer fires.
-func TestBefore_PipesInputOutputFieldThroughStages(t *testing.T) {
-	eng, reg, exec := newPreflightEnv(t)
-
-	exec.setReturnFn("render-obj", func(_, _ string, _ *task.Spec) interface{} {
-		return map[string]any{"path": "/var/run/relay.yaml", "size": 42}
-	})
-	render := makeOneShotSpec("render-obj")
-	if err := reg.Register(render); err != nil {
-		t.Fatalf("register render-obj: %v", err)
-	}
-
-	writer := &task.Spec{
-		ID:      "write-obj",
-		Name:    "write-obj",
-		Runtime: task.RuntimeDeno,
-		Trigger: task.TriggerConfig{Manual: true},
-		Params: task.Params{
-			{Name: "target", Required: true},
-		},
-		Enabled: true,
-	}
-	if err := reg.Register(writer); err != nil {
-		t.Fatalf("register writer: %v", err)
-	}
-
-	daemon := &task.Spec{
-		ID:      "d-obj",
-		Name:    "d-obj",
-		Runtime: task.RuntimeDocker,
-		Docker:  &task.DockerConfig{Image: "alpine"},
-		Trigger: task.TriggerConfig{
-			Daemon:  true,
-			Restart: "never",
-			Before: []task.BeforeEntry{
-				{Task: "render-obj"},
-				{
-					Task: "write-obj",
-					Overrides: &task.Overrides{
-						Params: task.ParamOverrides{
-							{Name: "target", Default: "${input.output.path}"},
-						},
-					},
-				},
-			},
-		},
-		Enabled: true,
-	}
-	if err := reg.Register(daemon); err != nil {
-		t.Fatalf("register daemon: %v", err)
-	}
-	exec.markDaemon("d-obj")
-
-	if err := eng.Register(daemon); err != nil {
-		t.Fatalf("eng.Register: %v", err)
-	}
-
-	var writeSpec *task.Spec
-	waitUntil(t, 5*time.Second, func() bool {
-		runs, _ := reg.ListRuns(context.Background(), "write-obj", 5)
-		for _, r := range runs {
-			if r.TriggerSource == registry.TriggerPreflight {
-				if s := exec.specForRun(r.ID); s != nil {
-					writeSpec = s
-					return true
-				}
-			}
-		}
-		return false
-	}, "write-obj stage never captured by executor")
-
-	var got string
-	for _, p := range writeSpec.Params {
-		if p.Name == "target" {
-			got = p.Default
-		}
-	}
-	if got != "/var/run/relay.yaml" {
-		t.Errorf("write-obj.target default = %q; want %q (output.field did not resolve)", got, "/var/run/relay.yaml")
-	}
-}
-
-// TestBefore_PipesInputOutputThroughStages verifies that stage[i+1]'s
-// overrides.params see the previous stage's return value substituted
-// for "${input.output}" at dispatch time. We assert by reading the
-// captured spec the second stage's executor received: its Params[content]
-// Default must equal the renderer's return value, not the literal token.
-func TestBefore_PipesInputOutputThroughStages(t *testing.T) {
-	eng, reg, exec := newPreflightEnv(t)
-
-	// Stage 1 returns the rendered string. The mid-pipeline contract
-	// is that this exact value lands on stage 2's content param.
-	exec.setReturnFn("render", func(_, _ string, _ *task.Spec) interface{} {
-		return "rendered-yaml"
-	})
-	render := makeOneShotSpec("render")
-	if err := reg.Register(render); err != nil {
-		t.Fatalf("register render: %v", err)
-	}
-
-	// Stage 2 (writer) declares a `content` param so the per-edge
-	// override can patch its Default with the upstream return value.
-	writer := &task.Spec{
-		ID:      "write",
-		Name:    "write",
-		Runtime: task.RuntimeDeno,
-		Trigger: task.TriggerConfig{Manual: true},
-		Params: task.Params{
-			{Name: "content", Required: true},
-		},
-		Enabled: true,
-	}
-	if err := reg.Register(writer); err != nil {
-		t.Fatalf("register writer: %v", err)
-	}
-
-	daemon := &task.Spec{
-		ID:      "d",
-		Name:    "d",
-		Runtime: task.RuntimeDocker,
-		Docker:  &task.DockerConfig{Image: "alpine"},
-		Trigger: task.TriggerConfig{
-			Daemon:  true,
-			Restart: "never",
-			Before: []task.BeforeEntry{
-				{Task: "render"},
-				{
-					Task: "write",
-					Overrides: &task.Overrides{
-						Params: task.ParamOverrides{
-							{Name: "content", Default: "${input.output}"},
-						},
-					},
-				},
-			},
-		},
-		Enabled: true,
-	}
-	if err := reg.Register(daemon); err != nil {
-		t.Fatalf("register daemon: %v", err)
-	}
-	exec.markDaemon("d")
-
-	if err := eng.Register(daemon); err != nil {
-		t.Fatalf("eng.Register: %v", err)
-	}
-
-	// Wait until the writer stage produced a captured spec.
-	var writeSpec *task.Spec
-	waitUntil(t, 5*time.Second, func() bool {
-		runs, _ := reg.ListRuns(context.Background(), "write", 5)
-		for _, r := range runs {
-			if r.TriggerSource == registry.TriggerPreflight {
-				if s := exec.specForRun(r.ID); s != nil {
-					writeSpec = s
-					return true
-				}
-			}
-		}
-		return false
-	}, "write stage never captured by executor")
-
-	var got string
-	for _, p := range writeSpec.Params {
-		if p.Name == "content" {
-			got = p.Default
-		}
-	}
-	if got != "rendered-yaml" {
-		t.Errorf("write.content default = %q; want %q (upstream output did not pipe through)", got, "rendered-yaml")
-	}
-}
-
-// TestBefore_FailureShortCircuits verifies that a failed stage halts
-// the pipeline: no descendant stage runs, and the daemon ends up in
-// DaemonPrereqFailed. Without short-circuit semantics, downstream stages
-// might run with stale upstream output (or an empty one) and leak
-// half-applied side effects.
-func TestBefore_FailureShortCircuits(t *testing.T) {
-	eng, reg, exec := newPreflightEnv(t)
-
-	exec.markFailing("stage-a")
-	var bCalled atomic.Bool
-	exec.setFn("stage-b", func(_, _ string) string {
-		bCalled.Store(true)
-		return registry.StatusSuccess
-	})
-
-	for _, id := range []string{"stage-a", "stage-b"} {
-		spec := makeOneShotSpec(id)
-		if err := reg.Register(spec); err != nil {
-			t.Fatalf("register %s: %v", id, err)
-		}
-	}
-
-	daemon := makeDaemonSpec("d", []string{"stage-a", "stage-b"})
-	if err := reg.Register(daemon); err != nil {
-		t.Fatalf("register daemon: %v", err)
-	}
-	exec.markDaemon("d")
-
-	if err := eng.Register(daemon); err != nil {
-		t.Fatalf("eng.Register: %v", err)
-	}
-
-	// Wait until daemon settles into PrereqFailed (the engine's terminal
-	// state after a failed preflight). 500ms is plenty given everything
-	// runs in-memory; pad to 2s to keep the test resilient on slow CI.
-	waitUntil(t, 2*time.Second, func() bool {
-		return eng.DaemonState("d") == DaemonPrereqFailed
-	}, "daemon never reached PrereqFailed")
-
-	if bCalled.Load() {
-		t.Error("stage-b ran despite stage-a failure; pipeline did not short-circuit")
-	}
-}
-
-// TestBefore_MidPipelineReFirePropagates verifies that re-firing an
-// intermediate stage replays its descendants with fresh
-// ${input.output} (and only its descendants — earlier stages are
-// untouched), then restarts the daemon.
-//
-// Without descendant propagation, the daemon would restart on the
-// re-fired stage's success but consume the writer's STALE last-known
-// output — defeating the point of mid-pipeline re-fire as a config-
-// rotation primitive.
-func TestBefore_MidPipelineReFirePropagates(t *testing.T) {
-	eng, reg, exec := newPreflightEnv(t)
-
-	var renderCalls atomic.Int32
-	exec.setReturnFn("render", func(_, _ string, _ *task.Spec) interface{} {
-		return fmt.Sprintf("render-run-%d", renderCalls.Add(1))
-	})
-	render := makeOneShotSpec("render")
-	if err := reg.Register(render); err != nil {
-		t.Fatalf("register render: %v", err)
-	}
-
-	// Capture each writer dispatch's content default so we can observe
-	// the propagated update.
-	var writerContents []string
-	var writerMu sync.Mutex
-	exec.setReturnFn("write", func(_, _ string, spec *task.Spec) interface{} {
-		writerMu.Lock()
-		defer writerMu.Unlock()
-		for _, p := range spec.Params {
-			if p.Name == "content" {
-				writerContents = append(writerContents, p.Default)
-			}
-		}
-		return nil
-	})
-	writer := &task.Spec{
-		ID:      "write",
-		Name:    "write",
-		Runtime: task.RuntimeDeno,
-		Trigger: task.TriggerConfig{Manual: true},
-		Params:  task.Params{{Name: "content", Required: true}},
-		Enabled: true,
-	}
-	if err := reg.Register(writer); err != nil {
-		t.Fatalf("register writer: %v", err)
-	}
-
-	daemon := &task.Spec{
-		ID:      "d",
-		Name:    "d",
-		Runtime: task.RuntimeDocker,
-		Docker:  &task.DockerConfig{Image: "alpine"},
-		Trigger: task.TriggerConfig{
-			Daemon:  true,
-			Restart: "never",
-			Before: []task.BeforeEntry{
-				{Task: "render"},
-				{
-					Task: "write",
-					Overrides: &task.Overrides{
-						Params: task.ParamOverrides{
-							{Name: "content", Default: "${input.output}"},
-						},
-					},
-				},
-			},
-		},
-		Enabled: true,
-	}
-	if err := reg.Register(daemon); err != nil {
-		t.Fatalf("register daemon: %v", err)
-	}
-	exec.markDaemon("d")
-
-	if err := eng.Register(daemon); err != nil {
-		t.Fatalf("eng.Register: %v", err)
-	}
-
-	// First boot: render-run-1 lands on writer.
-	waitUntil(t, 5*time.Second, func() bool {
-		writerMu.Lock()
-		defer writerMu.Unlock()
-		return len(writerContents) >= 1
-	}, "writer never received first content")
-	writerMu.Lock()
-	if writerContents[0] != "render-run-1" {
-		t.Errorf("writer[0] = %q; want render-run-1", writerContents[0])
-	}
-	writerMu.Unlock()
-	waitUntil(t, 5*time.Second, func() bool {
-		return eng.DaemonState("d") == DaemonRunning
-	}, "daemon never reached Running on first boot")
-
-	// Re-fire the renderer; the engine must replay the writer with the
-	// new render output, then restart the daemon.
-	if _, err := eng.FireManual(context.Background(), "render", nil); err != nil {
-		t.Fatalf("FireManual render: %v", err)
-	}
-
-	waitUntil(t, 10*time.Second, func() bool {
-		writerMu.Lock()
-		defer writerMu.Unlock()
-		return len(writerContents) >= 2
-	}, "writer never re-ran after render re-fire (no descendant propagation)")
-	writerMu.Lock()
-	got := writerContents[1]
-	writerMu.Unlock()
-	if got != "render-run-2" {
-		t.Errorf("writer[1] = %q; want render-run-2 (descendant did not see fresh upstream output)", got)
-	}
-
-	// Daemon must return to Running after the propagation completes.
-	if !waitForCond(10*time.Second, func() bool {
-		return eng.DaemonState("d") == DaemonRunning
-	}) {
-		t.Errorf("daemon never returned to Running after mid-pipeline re-fire; state=%v", eng.DaemonState("d"))
-	}
-}
-
-// TestBefore_MidPipelineReFireBailsOnUnregister verifies that the
-// propagation goroutine in propagateBeforeRerun bails out cleanly when
-// the daemon is unregistered mid-pipeline. Without the shutdown/
-// unregister guard, the propagation loop would keep firing descendant
-// stages and call startDaemonInternal on a daemon Unregister has
-// already purged from daemonSpecs — leaving stray daemonRuns entries
-// and racing the shutdown path.
-//
-// The test arranges a 2-descendant pipeline (render → write → finalize)
-// and blocks stage 2 ("write") on a channel so the propagation
-// goroutine is parked inside dispatchPipelineStage when we unregister
-// the daemon. After release + unregister, no third stage should be
-// dispatched, no DaemonRunning state should be set, and no stray
-// daemonRuns entry should remain.
-func TestBefore_MidPipelineReFireBailsOnUnregister(t *testing.T) {
-	eng, reg, exec := newPreflightEnv(t)
-
-	// Stage 1 (render): completes immediately, returns a token the
-	// later stages would (in the happy path) pipe through.
-	exec.setReturnFn("render", func(_, _ string, _ *task.Spec) interface{} {
-		return "render-out"
-	})
-	if err := reg.Register(makeOneShotSpec("render")); err != nil {
-		t.Fatalf("register render: %v", err)
-	}
-
-	// Stage 2 (write): blocks on a channel so the propagation
-	// goroutine is parked inside dispatchPipelineStage when we
-	// invoke Unregister below. write is the FIRST descendant of the
-	// re-fired render stage; we need it to actually start (so the
-	// "before the loop" guard cannot bail) but not finish until we
-	// release it.
-	writeStarted := make(chan struct{}, 1)
-	writeRelease := make(chan struct{})
-	var writeCalls atomic.Int32
-	exec.setFn("write", func(_, _ string) string {
-		if writeCalls.Add(1) == 1 {
-			select {
-			case writeStarted <- struct{}{}:
-			default:
-			}
-			<-writeRelease
-		}
-		return registry.StatusSuccess
-	})
-	if err := reg.Register(makeOneShotSpec("write")); err != nil {
-		t.Fatalf("register write: %v", err)
-	}
-
-	// Stage 3 (finalize): must NOT run after we unregister mid-loop.
-	// Count its dispatches so we can assert it stayed at zero
-	// post-release for the descendant that follows the blocked one.
-	var finalizeCalls atomic.Int32
-	exec.setFn("finalize", func(_, _ string) string {
-		finalizeCalls.Add(1)
-		return registry.StatusSuccess
-	})
-	if err := reg.Register(makeOneShotSpec("finalize")); err != nil {
-		t.Fatalf("register finalize: %v", err)
-	}
-
-	daemon := &task.Spec{
-		ID:      "d",
-		Name:    "d",
-		Runtime: task.RuntimeDocker,
-		Docker:  &task.DockerConfig{Image: "alpine"},
-		Trigger: task.TriggerConfig{
-			Daemon:  true,
-			Restart: "never",
-			Before: []task.BeforeEntry{
-				{Task: "render"},
-				{Task: "write"},
-				{Task: "finalize"},
-			},
-		},
-		Enabled: true,
-	}
-	if err := reg.Register(daemon); err != nil {
-		t.Fatalf("register daemon: %v", err)
-	}
-	exec.markDaemon("d")
-
-	if err := eng.Register(daemon); err != nil {
-		t.Fatalf("eng.Register: %v", err)
-	}
-
-	// First boot: writeCalls==1 (initial write during preflight) then
-	// the daemon comes up. Drain the first call so the next
-	// writeStarted signal corresponds to the propagation path.
-	select {
-	case <-writeStarted:
-	case <-time.After(5 * time.Second):
-		t.Fatalf("initial preflight write never started")
-	}
-	close(writeRelease) // unblock the first call; new writes will not block.
-	waitUntil(t, 5*time.Second, func() bool {
-		return eng.DaemonState("d") == DaemonRunning
-	}, "daemon never reached Running on first boot")
-
-	// Re-arm the block so the *second* (propagation) call to write
-	// parks inside Execute until we unregister.
-	writeRelease2 := make(chan struct{})
-	writeStarted2 := make(chan struct{}, 1)
-	exec.setFn("write", func(_, _ string) string {
-		select {
-		case writeStarted2 <- struct{}{}:
-		default:
-		}
-		<-writeRelease2
-		return registry.StatusSuccess
-	})
-
-	// Re-fire render — this kicks off propagation; write will block
-	// inside Execute, parking the propagation goroutine.
-	if _, err := eng.FireManual(context.Background(), "render", nil); err != nil {
-		t.Fatalf("FireManual render: %v", err)
-	}
-	select {
-	case <-writeStarted2:
-	case <-time.After(5 * time.Second):
-		t.Fatalf("propagation never reached the write stage")
-	}
-
-	// Now the propagation goroutine is parked inside
-	// dispatchPipelineStage. Unregister the daemon — Unregister
-	// purges daemonSpecs and KillRuns the live daemon run.
-	eng.Unregister("d")
-
-	// Capture finalize-call count and daemon-state snapshot at the
-	// moment of unregister so we can compare after the release.
-	finalizeBefore := finalizeCalls.Load()
-
-	// Release write — the propagation goroutine resumes, returns
-	// from dispatchPipelineStage, and (with the fix) sees the
-	// unregister + bails before dispatching finalize or restarting
-	// the daemon.
-	close(writeRelease2)
-
-	// Give the propagation goroutine a moment to observe the
-	// unregister and exit. We can't waitUntil on a negative ("never
-	// dispatches finalize") so we poll for a short window: if
-	// finalize stays at finalizeBefore and daemonRuns stays clean,
-	// the guard worked.
-	deadline := time.Now().Add(2 * time.Second)
-	for time.Now().Before(deadline) {
-		if finalizeCalls.Load() != finalizeBefore {
-			break
-		}
-		time.Sleep(10 * time.Millisecond)
-	}
-
-	if got := finalizeCalls.Load(); got != finalizeBefore {
-		t.Errorf("finalize was dispatched after unregister: calls=%d, want %d (propagation goroutine ignored the unregister)", got, finalizeBefore)
-	}
-
-	// No stray daemonRuns entry should remain after unregister: the
-	// propagation goroutine must NOT have re-fired startDaemonInternal,
-	// which is the load-bearing assertion (state Running may linger
-	// from the first boot because Unregister doesn't reset DaemonState
-	// — but startDaemonInternal would add a *new* daemonRuns entry,
-	// which Unregister has already drained).
-	eng.daemonMu.Lock()
-	leftover, hasRun := eng.daemonRuns["d"]
-	eng.daemonMu.Unlock()
-	if hasRun {
-		t.Errorf("daemonRuns has stray entry for unregistered daemon: runID=%q (propagation restarted the daemon after unregister)", leftover)
-	}
-}
-
-// TestRunPrereqs_BailsOnShutdown mirrors TestBefore_MidPipelineReFireBailsOnUnregister
-// but for the first-boot preflight path (runPrereqs, not propagateBeforeRerun).
-// Without the shutdown guard in runPrereqs, a long pipeline that spans an
-// engine shutdown window would keep dispatching descendant stages and
-// eventually push the daemon through to fireAsync on a torn-down engine.
-//
-// Setup: 3-stage before pipeline (stage1 → stage2 → stage3). Stage 1 blocks
-// inside Execute on a channel until released; while it's parked, we set the
-// engine's shutdownCtx to a cancelled context. After release, the loop's
-// top-of-iteration guard (or the engine context cancellation in
-// dispatchPipelineStage's WaitRun) must bail before dispatching stage 3,
-// and the daemon must never transition to DaemonRunning.
-func TestRunPrereqs_BailsOnShutdown(t *testing.T) {
-	eng, reg, exec := newPreflightEnv(t)
-
-	// Stage 1: blocks on a channel so the preflight loop is parked
-	// inside dispatchPipelineStage when we cancel the engine context.
-	stage1Started := make(chan struct{}, 1)
-	stage1Release := make(chan struct{})
-	exec.setFn("stage1", func(_, _ string) string {
-		select {
-		case stage1Started <- struct{}{}:
-		default:
-		}
-		<-stage1Release
-		return registry.StatusSuccess
-	})
-	if err := reg.Register(makeOneShotSpec("stage1")); err != nil {
-		t.Fatalf("register stage1: %v", err)
-	}
-
-	// Stage 2 / 3: count their dispatches so we can assert they stayed
-	// at zero after the engine context is cancelled mid-pipeline.
-	var stage2Calls atomic.Int32
-	exec.setFn("stage2", func(_, _ string) string {
-		stage2Calls.Add(1)
-		return registry.StatusSuccess
-	})
-	if err := reg.Register(makeOneShotSpec("stage2")); err != nil {
-		t.Fatalf("register stage2: %v", err)
-	}
-	var stage3Calls atomic.Int32
-	exec.setFn("stage3", func(_, _ string) string {
-		stage3Calls.Add(1)
-		return registry.StatusSuccess
-	})
-	if err := reg.Register(makeOneShotSpec("stage3")); err != nil {
-		t.Fatalf("register stage3: %v", err)
-	}
-
-	daemon := &task.Spec{
-		ID:      "d",
-		Name:    "d",
-		Runtime: task.RuntimeDocker,
-		Docker:  &task.DockerConfig{Image: "alpine"},
-		Trigger: task.TriggerConfig{
-			Daemon:  true,
-			Restart: "never",
-			Before: []task.BeforeEntry{
-				{Task: "stage1"},
-				{Task: "stage2"},
-				{Task: "stage3"},
-			},
-		},
-		Enabled: true,
-	}
-	if err := reg.Register(daemon); err != nil {
-		t.Fatalf("register daemon: %v", err)
-	}
-	exec.markDaemon("d")
-
-	if err := eng.Register(daemon); err != nil {
-		t.Fatalf("eng.Register: %v", err)
-	}
-
-	// Wait for stage 1 to actually be inside Execute — at this point
-	// the preflight loop is parked and the engine has not yet
-	// transitioned past stage 1.
-	select {
-	case <-stage1Started:
-	case <-time.After(5 * time.Second):
-		t.Fatalf("preflight stage1 never started")
-	}
-
-	// Install a cancelled shutdown context so isShuttingDown() returns
-	// true. The top-of-iteration guard in runPrereqs must observe this
-	// before dispatching stage 2.
-	shutCtx, cancel := context.WithCancel(context.Background())
-	cancel()
-	eng.shutdownMu.Lock()
-	eng.shutdownCtx = shutCtx
-	eng.shutdownMu.Unlock()
-
-	// Release stage 1 — the preflight loop resumes, sees the
-	// cancelled engine context at the top of the next iteration, and
-	// returns nil. Stages 2 and 3 must never run; the daemon must
-	// never transition to Running.
-	close(stage1Release)
-
-	// Poll briefly to give the preflight goroutine time to observe the
-	// shutdown and exit. We can't waitUntil on a negative ("never
-	// dispatches stage2/3"), so we wait a short window.
-	deadline := time.Now().Add(2 * time.Second)
-	for time.Now().Before(deadline) {
-		if stage2Calls.Load() > 0 || stage3Calls.Load() > 0 {
-			break
-		}
-		time.Sleep(10 * time.Millisecond)
-	}
-
-	if got := stage2Calls.Load(); got != 0 {
-		t.Errorf("stage2 dispatched after shutdown: calls=%d, want 0", got)
-	}
-	if got := stage3Calls.Load(); got != 0 {
-		t.Errorf("stage3 dispatched after shutdown: calls=%d, want 0", got)
-	}
-
-	// Daemon must NOT have reached Running — runPrereqs returning nil
-	// on shutdown still falls through to fireAsync, BUT fireAsync
-	// receives the same cancelled context… actually no, fireAsync is
-	// called with context.Background() from startDaemonInternal's
-	// post-prereq path. The load-bearing assertion is therefore on
-	// stage2/3 above: if those stayed at zero, runPrereqs bailed.
-	// We additionally assert the daemon's state is not Running to
-	// confirm the preflight short-circuit didn't accidentally
-	// short-circuit into success.
-	//
-	// State note: startDaemonInternal currently transitions through
-	// PrereqRunning → (skip fireAsync on shutdown via runPrereqs nil)
-	// → falls through to fireAsync(context.Background()). Because
-	// fireAsync's context is NOT shutdown-bound today, the daemon
-	// could theoretically start. The test asserts the load-bearing
-	// guarantee — stages 2/3 stayed at zero — which is the only
-	// behavior runPrereqs itself controls. A future tightening of
-	// fireAsync's context handling would make the daemon-state
-	// assertion meaningful too; for now it is informational.
-	state := eng.DaemonState("d")
-	if state == DaemonRunning {
-		// Not a hard failure (see comment above), but log for context.
-		t.Logf("note: daemon reached Running despite shutdown — fireAsync uses context.Background; runPrereqs guard still worked (stage2/3 zero)")
-	}
-}
-
-// TestRegister_RejectsDuplicateBeforeTask verifies the duplicate-detection
-// added to validateBeforeRefs. propagateBeforeRerun's startIdx lookup
-// breaks at the first match, so if the same task appears twice in before:
-// only the descendants of the first occurrence get replayed on re-fire.
-// The validator must reject the spec at registration so operators see the
-// ambiguity at config-load time rather than discover it via missed re-fires.
-func TestRegister_RejectsDuplicateBeforeTask(t *testing.T) {
-	e := newTestEnv(t)
-	prereq := &task.Spec{
-		ID:      "render",
-		Name:    "render",
-		Runtime: task.RuntimeDeno,
-		Trigger: task.TriggerConfig{Manual: true},
-		Enabled: true,
-	}
-	if err := e.reg.Register(prereq); err != nil {
-		t.Fatalf("seed reg.Register: %v", err)
-	}
-	daemon := &task.Spec{
-		ID:      "d",
-		Name:    "d",
-		Runtime: task.RuntimeDocker,
-		Docker:  &task.DockerConfig{Image: "alpine"},
-		Trigger: task.TriggerConfig{
-			Daemon: true,
-			Before: []task.BeforeEntry{
-				{Task: "render"},
-				{Task: "render"},
-			},
-		},
-		Enabled: true,
-	}
-	err := e.engine.Register(daemon)
-	if err == nil {
-		t.Fatalf("expected error rejecting duplicate task in before, got nil")
-	}
-	if !strings.Contains(err.Error(), "multiple times") {
-		t.Errorf("expected error mentioning duplication, got: %v", err)
-	}
-}
-
-// TestStartDaemonInternal_FireAsyncFailureAfterPreflight verifies that when
-// startDaemonInternal's preflight stage succeeds (or is deliberately skipped
-// via skipPrereqs=true from the mid-pipeline re-fire path) but the
-// subsequent fireAsync fails, the daemon transitions to
-// DaemonFailedAfterPreflight — NOT to the generic DaemonStopped.
-//
-// Operators looking at the WebUI must be able to tell "preflight is fine,
-// daemon body broke" apart from "deliberately stopped / never started".
-// See issue #318 for the motivating scenario (Doppler-rotated relay-server
-// preflight succeeds, but the daemon's docker run fails on a port-already-
-// bound or image-pull error).
-//
-// We force the fireAsync failure by closing the underlying DB so
-// StartRunWithID — the first thing startRun does inside fireAsync — fails.
-// That's a low-fidelity stand-in for the real failure modes (binary
-// missing, port bound) but exercises the same code path: fireAsync returns
-// an error, startDaemonInternal hits its post-preflight error branch, and
-// the state assignment is the system-under-test.
-func TestStartDaemonInternal_FireAsyncFailureAfterPreflight(t *testing.T) {
-	// Build the env inline so we own the DB and can close it
-	// out from under the engine to force fireAsync to fail.
-	d, err := db.Open(db.Config{Type: "sqlite", Path: ":memory:"})
-	if err != nil {
-		t.Fatalf("open db: %v", err)
-	}
-	reg := registry.New(d)
-	exec := newPreflightExec(reg)
-	eng := New(reg, exec, zap.NewNop())
-	eng.RegisterExecutor(task.RuntimeDocker, exec)
-
-	// No preflight stages — keeps the test focused on the post-preflight
-	// fireAsync failure. With skipPrereqs=false and an empty Before list,
-	// startDaemonInternal goes straight to the fireAsync call, which is
-	// the exact branch the issue is about.
-	daemon := &task.Spec{
-		ID:      "d",
-		Name:    "d",
-		Runtime: task.RuntimeDocker,
-		Docker:  &task.DockerConfig{Image: "alpine"},
-		Trigger: task.TriggerConfig{Daemon: true, Restart: "never"},
-		Enabled: true,
-	}
-	if err := reg.Register(daemon); err != nil {
-		t.Fatalf("reg.Register: %v", err)
-	}
-
-	// Close the DB so any subsequent registry write (StartRunWithID, in
-	// our case) fails. Mirrors what would happen if disk space ran out or
-	// the DB connection was severed mid-operation.
-	if err := d.Close(); err != nil {
-		t.Fatalf("db.Close: %v", err)
-	}
-
-	// Drive startDaemonInternal directly with skipPrereqs=true — mirrors
-	// the call path used by propagateBeforeRerun (the issue's motivating
-	// scenario). Synchronous: fireAsync returns the error inline, so we
-	// can assert the state immediately without polling.
-	eng.startDaemonInternal(daemon, true)
-
-	if got := eng.DaemonState("d"); got != DaemonFailedAfterPreflight {
-		t.Fatalf("DaemonState = %q, want %q (must distinguish post-preflight failure from deliberately-stopped)",
-			got, DaemonFailedAfterPreflight)
-	}
-
-	// Sanity check: a daemon that was never started and never failed
-	// must NOT appear as failed_after_preflight — that would be a false
-	// positive on every fresh task.
-	if got := eng.DaemonState("never-touched"); got != DaemonStopped {
-		t.Errorf("DaemonState(unknown) = %q, want %q", got, DaemonStopped)
-	}
-}
-
-// TestStartDaemonInternal_FireAsyncFailureFirstBoot covers the sibling
-// branch of TestStartDaemonInternal_FireAsyncFailureAfterPreflight:
-// startDaemonInternal called with skipPrereqs=false (the first-boot path
-// taken by Engine.startDaemon when a daemon is initially scheduled). With
-// an empty trigger.before list the prereq block is skipped and execution
-// falls straight through to fireAsync — same error branch, different
-// caller — so the engine MUST still record DaemonFailedAfterPreflight
-// rather than DaemonStopped.
-//
-// The engine change in issue #318 applies to BOTH callers
-// (propagateBeforeRerun's skipPrereqs=true and the first-boot
-// skipPrereqs=false); without this sibling test, regressions on the
-// first-boot branch would pass CI.
-//
-// Same DB-close stand-in as the skipPrereqs=true sibling.
-func TestStartDaemonInternal_FireAsyncFailureFirstBoot(t *testing.T) {
-	d, err := db.Open(db.Config{Type: "sqlite", Path: ":memory:"})
-	if err != nil {
-		t.Fatalf("open db: %v", err)
-	}
-	reg := registry.New(d)
-	exec := newPreflightExec(reg)
-	eng := New(reg, exec, zap.NewNop())
-	eng.RegisterExecutor(task.RuntimeDocker, exec)
-
-	// Empty Before list + skipPrereqs=false means startDaemonInternal
-	// skips its prereq block via the len(spec.Trigger.Before) > 0 guard
-	// and goes straight to fireAsync — the exact branch this sibling
-	// covers. Adding stages here would couple the test to runPrereqs's
-	// behavior, which is not the system under test.
-	daemon := &task.Spec{
-		ID:      "d",
-		Name:    "d",
-		Runtime: task.RuntimeDocker,
-		Docker:  &task.DockerConfig{Image: "alpine"},
-		Trigger: task.TriggerConfig{Daemon: true, Restart: "never"},
-		Enabled: true,
-	}
-	if err := reg.Register(daemon); err != nil {
-		t.Fatalf("reg.Register: %v", err)
-	}
-
-	if err := d.Close(); err != nil {
-		t.Fatalf("db.Close: %v", err)
-	}
-
-	// skipPrereqs=false — the first-boot branch. fireAsync still fails
-	// (DB closed), so the state assignment must match the sibling test.
-	eng.startDaemonInternal(daemon, false)
-
-	if got := eng.DaemonState("d"); got != DaemonFailedAfterPreflight {
-		t.Fatalf("DaemonState = %q, want %q (first-boot path must record post-preflight failure identically to the mid-pipeline re-fire path)",
-			got, DaemonFailedAfterPreflight)
-	}
-}
-
-// finishedRunEnv stages a daemon spec + a finished registry run so the
-// onDaemonRunFinished tests can call the hook directly without going
-// through the full fireAsync → daemon-execute path. Keeping the setup
-// inline (rather than reusing preflightExec/newPreflightEnv) avoids
-// coupling the test to the executor's daemon-block semantics — these
-// tests are about the post-run state-transition decision only.
-func finishedRunEnv(t *testing.T, restartPolicy string, status string) (*Engine, *task.Spec, string) {
-	t.Helper()
-	d, err := db.Open(db.Config{Type: "sqlite", Path: ":memory:"})
-	if err != nil {
-		t.Fatalf("open db: %v", err)
-	}
-	t.Cleanup(func() { d.Close() })
-	reg := registry.New(d)
-	exec := newPreflightExec(reg)
-	eng := New(reg, exec, zap.NewNop())
-	eng.RegisterExecutor(task.RuntimeDocker, exec)
-
-	spec := &task.Spec{
-		ID:      "d",
-		Name:    "d",
-		Runtime: task.RuntimeDocker,
-		Docker:  &task.DockerConfig{Image: "alpine"},
-		Trigger: task.TriggerConfig{Daemon: true, Restart: restartPolicy},
-		Enabled: true,
-	}
-	if err := reg.Register(spec); err != nil {
-		t.Fatalf("reg.Register: %v", err)
-	}
-	// Mark the daemon as still-registered with the engine so the
-	// onDaemonRunFinished early-return guard doesn't bail out.
-	eng.daemonMu.Lock()
-	eng.daemonSpecs[spec.ID] = spec
-	eng.daemonMu.Unlock()
-
-	// Seed a run row in the registry and finalize it with the requested
-	// terminal status — onDaemonRunFinished looks the run up via GetRun
-	// to decide whether to restart.
-	ctx := context.Background()
-	runID, err := reg.StartRun(ctx, spec.ID, "")
-	if err != nil {
-		t.Fatalf("StartRun: %v", err)
-	}
-	if err := reg.FinishRun(ctx, runID, status); err != nil {
-		t.Fatalf("FinishRun: %v", err)
-	}
-
-	// Pre-seed daemonStates with DaemonRunning so we can assert the
-	// transition actually fires (rather than just observing the default
-	// "stopped" fall-through that DaemonState() returns for absent keys).
-	// This mirrors reality: the WebUI's stale-Running pill described in
-	// issue #325 only appears because the engine successfully set
-	// DaemonRunning before the body crashed.
-	eng.setDaemonState(spec.ID, DaemonRunning)
-	return eng, spec, runID
-}
-
-// TestOnDaemonRunFinished_NoRestartFailure_TransitionsToCrashed exercises
-// the issue #325 fix: a daemon whose body exits with status=failure under
-// `restart: never` must transition to DaemonCrashed so the WebUI shows a
-// distinct pill (operator attention required) rather than a stale Running
-// indicator. Mirror test for the success path lives in the sibling
-// _TransitionsToStopped test below.
-func TestOnDaemonRunFinished_NoRestartFailure_TransitionsToCrashed(t *testing.T) {
-	eng, spec, runID := finishedRunEnv(t, "never", registry.StatusFailure)
-	eng.onDaemonRunFinished(spec, runID)
-	if got := eng.DaemonState(spec.ID); got != DaemonCrashed {
-		t.Fatalf("DaemonState = %q, want %q (non-success exit + no restart must surface as Crashed, not stale Running)",
-			got, DaemonCrashed)
-	}
-}
-
-// TestOnDaemonRunFinished_NoRestartSuccess_TransitionsToStopped covers
-// the clean-exit half of issue #325's status-aware split: a daemon body
-// that exits successfully under `restart: never` (e.g. a long-running
-// migration that completes) is NOT a crash — it's a clean shutdown, and
-// the operator should see Stopped, not Crashed.
-func TestOnDaemonRunFinished_NoRestartSuccess_TransitionsToStopped(t *testing.T) {
-	eng, spec, runID := finishedRunEnv(t, "never", registry.StatusSuccess)
-	eng.onDaemonRunFinished(spec, runID)
-	if got := eng.DaemonState(spec.ID); got != DaemonStopped {
-		t.Fatalf("DaemonState = %q, want %q (clean exit + no restart must surface as Stopped, not Crashed)",
-			got, DaemonStopped)
-	}
-}
-
-// TestOnDaemonRunFinished_OnFailurePolicy_SuccessExit_TransitionsToStopped
-// covers the second no-restart branch: restart=on-failure with a
-// success exit. The engine won't restart (no failure to react to), so
-// the daemon is in the same terminal Stopped state as the
-// restart=never+success case.
-func TestOnDaemonRunFinished_OnFailurePolicy_SuccessExit_TransitionsToStopped(t *testing.T) {
-	eng, spec, runID := finishedRunEnv(t, "on-failure", registry.StatusSuccess)
-	eng.onDaemonRunFinished(spec, runID)
-	if got := eng.DaemonState(spec.ID); got != DaemonStopped {
-		t.Fatalf("DaemonState = %q, want %q (on-failure + clean exit must surface as Stopped — no failure to restart, no crash to surface)",
-			got, DaemonStopped)
-	}
-}
-
-// TestOnDaemonRunFinished_Cancelled_RestartNever_TransitionsToStopped
-// pins the fix for issue #332: PR #329's status-aware no-restart split
-// (Stopped vs Crashed) didn't cover the StatusCancelled path because
-// cancellation hits an earlier early-return in onDaemonRunFinished. A
-// daemon killed via the per-run kill button under `restart: never` was
-// left in the DaemonRunning state the engine had set when it started —
-// the same stale-Running-pill UX bug #329 fixed for failure/success
-// exits. Operator-initiated cancellation is semantically a clean stop,
-// so it routes into DaemonStopped (Option A from #332).
-func TestOnDaemonRunFinished_Cancelled_RestartNever_TransitionsToStopped(t *testing.T) {
-	eng, spec, runID := finishedRunEnv(t, "never", registry.StatusCancelled)
-	eng.onDaemonRunFinished(spec, runID)
-	if got := eng.DaemonState(spec.ID); got != DaemonStopped {
-		t.Fatalf("DaemonState = %q, want %q (operator-initiated cancellation must surface as Stopped, not stale Running)",
-			got, DaemonStopped)
-	}
-}
-
-// TestOnDaemonRunFinished_Cancelled_RestartAlways_TransitionsToStopped
-// locks in the "regardless of restart policy" half of the #332
-// contract: cancellation is operator intent, so even when the spec
-// says `restart: always` the engine must NOT loop the daemon back to
-// Running — it must land in Stopped. Without this guard, a kill from
-// the WebUI on a restart-always daemon would just respawn it, making
-// the kill button feel broken.
-func TestOnDaemonRunFinished_Cancelled_RestartAlways_TransitionsToStopped(t *testing.T) {
-	eng, spec, runID := finishedRunEnv(t, "always", registry.StatusCancelled)
-	eng.onDaemonRunFinished(spec, runID)
-	if got := eng.DaemonState(spec.ID); got != DaemonStopped {
-		t.Fatalf("DaemonState = %q, want %q (operator cancellation overrides restart=always — kill button must stick)",
-			got, DaemonStopped)
-	}
-}
-
-// TestOnDaemonRunFinished_Cancelled_RestartOnFailure_TransitionsToStopped
-// covers the third restart policy. On-failure normally would NOT
-// restart a cancellation (it's not a failure), so the test mostly
-// confirms the StatusCancelled branch doesn't accidentally leak into
-// the failure-classification path and trigger a spurious restart or
-// Crashed transition.
-func TestOnDaemonRunFinished_Cancelled_RestartOnFailure_TransitionsToStopped(t *testing.T) {
-	eng, spec, runID := finishedRunEnv(t, "on-failure", registry.StatusCancelled)
-	eng.onDaemonRunFinished(spec, runID)
-	if got := eng.DaemonState(spec.ID); got != DaemonStopped {
-		t.Fatalf("DaemonState = %q, want %q (cancellation under restart=on-failure must surface as Stopped, not Crashed)",
-			got, DaemonStopped)
-	}
-}
